Rearrange folder structure for Library Manager compatibility
As stated in the Library Manager FAQ:
>If your library is stored in a subfolder, your repo is not OK and we
can't include your library in the list.

This change will also allow the .ZIP download of the library available
on GitHub to be installed via Sketch > Include Library > Add .ZIP
Library...

I placed the library source files directly in the root folder to allow
backwards compatibility with Arduino IDE 1.0.x. If that's not considered
desirable then they can be placed in a subfolder named src in accordance
with the 1.5 Library Format specification:

https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#layout-of-folders-and-files

I merged the two .gitignore files.